refactor timeslot components
